Devdocs: Add sphinx_autodoc_typehints
One of the big innovations of Python 3.5 was support for type hints.
This extension allows Sphinx to collect the types of functions'
parameters automatically from the function definition, obviating the
need to note the type of a parameter:

:param int xyz: becomes just :param xyz:

Type hints remain optional, but where provided they can cut down
somewhat on developer work.